django-backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
django-environ>=0.11.2

# GraphQL
strawberry-graphql-django>=0.38.0

# Stellar SDK
stellar-sdk>=10.0.0

# Database
psycopg2-binary>=2.9.9

# Task Queue
celery>=5.3.4
redis>=5.0.1

# Logging & Error Tracking
python-json-logger>=2.0.0
sentry-sdk[django,celery]>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
jsonschema>=4.0.0

# Development
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
responses>=0.24.0
black>=23.12.0
ruff>=0.1.8
pytest-mock>=3.12.0
